Adjust teacher related endpoints to use reference instead of id

+++ LibraryManager.Domain/Dtos/Students/UpdateStudentTeacherDto.cs
@@ -3,6 +3,6 @@
     public class UpdateStudentTeacherDto
     {
         public long StudentId { get; set; }
-        public long TeacherId { get; set; }
+        public int TeacherId { get; set; }
     }
 }
